redesign send amount
