+++ src/locale/locales/es/common.json
@@ -53,10 +53,10 @@
     "send_to": "Enviar a",
     "address": "Dirección",
     "scanned_address": "Dirección escaneada: {{address}}",
-    "transaction_fee": "Tarifa de la transacción: {{fee}} CKB",
+    "transaction_fee": "Tarifa de la transacción: {{fee}} {{token}}",
     "write_a_message": "Escribe un mensaje",
     "enter_amount": "Añade cantidad",
-    "minimum_transaction_amount_text": "La cantidad mínima requerida es de {{amount}} CKB",
+    "minimum_transaction_amount_text": "La cantidad mínima requerida es de {{amount}} {{token}}",
     "insufficient_balance": "Balance insuficiente",
     "modify_default_fee": "Modifica la tarifa por defecto",
     "general_settings": "Ajustes generales",
