+++ test/unit/src/module/transaction/screen/SendSetAmountScreen.spec.tsx
@@ -34,8 +34,8 @@ describe("SendAmountAndMessageScreen tests", () => {
     test("Renders correctly", async () => {
         const screen = render(<SendSetAmountScreen />);
         await waitFor(() => expect(screen.getByPlaceholderText(translate("enter_amount"))).toBeDefined());
-        expect(screen.getByText("CKB")).toBeDefined();
-        expect(screen.getByText(translate("transaction_fee", { fee: "0.001" }))).toBeDefined();
+        expect(screen.getByText(config.tokenName)).toBeDefined();
+        expect(screen.getByText(translate("transaction_fee", { fee: "0.001", token: config.tokenName }))).toBeDefined();
         expect(screen.getByPlaceholderText(translate("write_a_message"))).toBeDefined();
         expect(screen.getByText(translate("next"))).toBeDefined();
     });
